app/WEBPAGE/AI_Assistant/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
flask>=2.3.0
requests>=2.31.0